Add initial support for subscriber mode:
When we use any of SUBSCRIBE commands, we need to restrict the
commands we can use after. Only other SUBSCRIBE, or UNSUBSCRIBE commands
are valid.

We added a "mode", is_subscriber, that default to 0 and will be true after
we subscribe something. When is_subscriber is active, invalid commands
will result in an exception.

Signed-off-by: Pedro Melo <melo@simplicidade.org>