Make sure we catch dead servers and timeouts:
As pointed out by @trinitum at #1, when the servers goes away, either
from a client timeout or a dead server, the write() might succeed, so the
auto-reconnect code would not detect that.

We now do a non-blocking read before the write() to detect those
situations and reconnect.

Notice that read errors while reading a response (as in a incomplete
response) will still trigger an exception, as this most likelly mean
lost data.

#1: https://github.com/melo/perl-redis/issues/1#issuecomment-3444989

Signed-off-by: Pedro Melo <melo@simplicidade.org>