Fixing inconsistency with binary strings (called "the unicode bug").
In perl, string which are identical (eq operator) should behave as same strings.
Fixing inconsistency when binary strings with UTF-8 flag set and flag clear
behave different way.

Intention of the module to consider all input and output data as binary data,
proper way to implement this does not involve call to _utf8_off, instead
it just works in pure-perl.

If data with wide characters is sent (i.e. characters with code > 255), it's can't
be downgaded, thus cannot be considered as binary string, thus exception
"Wide characters in syswrite" is thrown.

Fixes #95