Fix for "Unexpected error condition 104/linux"
If Redis connection is routed thorugh an intermediate proxy (e.g. HAProxy)
it might be forcefully reset due to internal proxy timeout settings.

This patch attempts to fix that by making sure that we treat ECONNRESET
error code on all platforms, not just freebsd. Effectively this change
would make Redis library to reconnect in the event of RST packet
received.